no message
Match-id-cc0338f6801bd056d7dc9dd5f94f69a37bed5f16

+++ trustedcore_hm/kernel/libgic/gic_v3.c
@@ -146,7 +146,7 @@ static void gic_set_irq_priority(uint32_t irq, uint8_t prio)
     /* clear old priority */
     new_prio = old_prio & ~(INT_PRIO_MASK << irq_bit);
     /* set new priorty */
-    new_prio |= (prio << irq_bit);
+    new_prio |= ((uint32_t)prio << irq_bit);
     if (irq >= IRQ_SPI_START) {
         write32_dist_reg(GICD_IPRIORITYRn + irq_base, new_prio);
         gic_wait_for_rwp();
